requirements.txt
cdsapi==0.7.0
streamlit==1.36.0
loguru==0.7.2
PyYAML==6.0.1
matplotlib==3.9.1
Cartopy==0.23.0
googlemaps==4.10.0
transformers==4.42.4
torch==2.3.1
xarray==2024.6.0